Move layer file info logging to containerize()
